+++ types/config.d.ts
@@ -265,7 +265,7 @@ export namespace Config {
      * User interface language
      * Same as {@link IAppearance.lang}
      */
-    export type TLang = "en_US" | "es_ES" | "fr_FR" | "zh_CHT" | "zh_CN";
+    export type TLang = "en_US" | "es_ES" | "fr_FR" | "zh_CHT" | "zh_CN" | "ja_JP";
 
     /**
      * SiYuan bazaar related configuration
@@ -282,9 +282,42 @@ export namespace Config {
     }
 
     /**
+     * SiYuan editor markdown related configuration
+     */
+    interface IMarkdown {
+        /**
+         * Whether to enable the inline superscript
+         */
+        inlineSup: boolean;
+        /**
+         * Whether to enable the inline subscript
+         */
+        inlineSub: boolean;
+        /**
+         * Whether to enable the inline tag
+         */
+        inlineTag: boolean;
+        /**
+         * Whether to enable the inline math
+         */
+        inlineMath: boolean;
+    }
+
+    /**
      * SiYuan editor related configuration
      */
     export interface IEditor {
+
+        /**
+         * Whether to allow to execute javascript in the HTML block
+         */
+        allowHTMLBLockScript: boolean;
+
+        /**
+         * Markdown configuration
+         */
+        markdown: IMarkdown;
+
         /**
          * The default number of backlinks to expand
          */
@@ -538,6 +571,8 @@ export namespace Config {
          * The storage path of the new document created using block references
          */
         refCreateSavePath: string;
+        refCreateSaveBox: string;
+        docCreateSaveBox: string;
         /**
          * Close the secondary confirmation when deleting a document
          */
@@ -1323,8 +1358,11 @@ export namespace Config {
         appDir: string;
         /**
          * Boot automatically
+         * - `0`: Do not boot automatically
+         * - `1`: Boot automatically
+         * - `2`: Boot automatically + Minimize UI
          */
-        autoLaunch: boolean;
+        autoLaunch2: number;
         /**
          * The absolute path of the `conf` directory of the current workspace
          */