+++ CHANGELOG.md
@@ -1,6 +1,20 @@
 # Changelog
 
-## v0.9.9 2024-04
+## v1.0.3 2024-06
+
+## v1.0.2 2024-06-04
+
+* [Add `dialog` object to `Setting` property](https://github.com/siyuan-note/petal/issues/28)
+
+## v0.9.9 2024-05-14
+
+* Update Config.d.ts
+* [Update plugin `loaded-protyle-dynamic` parameter positon to position](https://github.com/siyuan-note/siyuan/pull/11298)
+
+## v0.9.8 2024-04-30
+
+* [Add `editors` to the plugin's `Custom` class](https://github.com/siyuan-note/siyuan/issues/11072)
+* [Add `direction` to plugin method `Setting.addItem`](https://github.com/siyuan-note/siyuan/issues/11183)
 
 ## v0.9.7 2024-04-02